feature/tokenless-service-refresh (#32)
* Updated API docs

* Added test case for tokenless refresh when authenticated

* Allows services to be refreshed without a token when authenticated

* Updated test cases

* Specifying api guard since auth:api middleware not used

+++ app/Http/Controllers/Core/V1/Service/RefreshController.php
@@ -27,7 +27,9 @@ class RefreshController extends Controller
             $service->update(['last_modified_at' => Date::now()]);
 
             // Delete the token used.
-            ServiceRefreshToken::query()->findOrFail($request->token)->delete();
+            if ($request->has('token')) {
+                ServiceRefreshToken::query()->findOrFail($request->token)->delete();
+            }
 
             event(EndpointHit::onUpdate($request, "Refreshed service [{$service->id}]", $service));
